+++ backend/src/auth/auth.controller.ts
@@ -14,7 +14,8 @@ import { User } from '../users/decorators/user.decorator';
 import { AuthService } from './auth.service';
 import { OauthCookies } from './decorators/oauth-data.decorator';
 import { SkipAuthGuard } from './decorators/public.decorator';
-import { OauthAuthorizeDto, OauthDto } from './dtos/oauth.dto';
+import { OauthAuthorizeDto } from './dtos/oauth-authorize.dto';
+import { OauthDto } from './dtos/oauth.dto';
 import { OauthDataGuard } from './guards/oauth-data.guard';
 import { RefreshTokenGuard } from './guards/refresh-token.guard';
 import { CookieInterceptor } from './interceptors/cookie.interceptor';
