+++ backend/src/auth/decorators/oauth-data.decorator.ts
@@ -1,5 +1,5 @@
 import { ExecutionContext, createParamDecorator } from '@nestjs/common';
-import { OauthData } from '../types/auth.type';
+import { OauthData } from '../types/oauth-data.type';
 
 export const OauthCookies = createParamDecorator((data: never, context: ExecutionContext): OauthData => {
     const request = context.switchToHttp().getRequest();
