+++ backend/src/auth/interceptors/cookie.interceptor.ts
@@ -4,7 +4,8 @@ import { CookieOptions, Response } from 'express';
 import { Observable, map } from 'rxjs';
 import { WinstonLoggerService } from 'src/common/logger/winstonLogger.service';
 import { TOKEN_LIFETIME_MAP } from '../token/token.service';
-import { AuthData, OauthData } from '../types/auth.type';
+import { AuthData } from '../types/auth-data.type';
+import { OauthData } from '../types/oauth-data.type';
 
 @Injectable()
 export class CookieInterceptor implements NestInterceptor {
