+++ backend/src/auth/types/auth-data.type.ts
@@ -1,0 +1,4 @@
+export interface AuthData {
+    accessToken: string;
+    refreshToken: string;
+}
