+++ backend/src/auth/types/oauth-authorize-data.type.ts
@@ -1,0 +1,6 @@
+import { OauthProvider } from './oauth-provider.type';
+
+export interface OauthAuthorizeData {
+    authorizeCode: string;
+    provider: OauthProvider;
+}
