+++ backend/src/auth/types/oauth-data.type.ts
@@ -1,0 +1,7 @@
+import { OauthProvider } from './oauth-provider.type';
+
+export interface OauthData {
+    oauthAccessToken: string;
+    oauthRefreshToken: string;
+    provider: OauthProvider;
+}
