+++ backend/src/auth/types/oauth-provider.type.ts
@@ -1,0 +1,3 @@
+export const OAUTH_PROVIDERS = ['google', 'kakao', 'naver'] as const;
+
+export type OauthProvider = (typeof OAUTH_PROVIDERS)[number];
